Factor convenience functions out of main printer implementation
